rename quote provider
